+++ CHANGELOG.md
@@ -3,6 +3,91 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.4](https://github.com/baloise/ui-library/compare/v2.3.3...v2.3.4) (2021-01-04)
+
+
+### Bug Fixes
+
+* **navbar:** fix navbar dropdown colors + add example ([64197bc](https://github.com/baloise/ui-library/commit/64197bc5cd50bd21cac1f717a29523ac2d8f51d3))
+
+
+
+## 2.1.1 (2020-12-23)
+
+
+
+
+
+## [2.3.3](https://github.com/baloise/ui-library/compare/v2.3.2...v2.3.3) (2021-01-04)
+
+
+
+## [2.3.1](https://github.com/baloise/ui-library/compare/v2.3.0...v2.3.1) (2021-01-04)
+
+
+
+# 2.3.0 (2021-01-04)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [2.3.2](https://github.com/baloise/ui-library/compare/v2.3.1...v2.3.2) (2021-01-04)
+
+
+
+# 2.3.0 (2021-01-04)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [2.3.1](https://github.com/baloise/ui-library/compare/v2.3.0...v2.3.1) (2021-01-04)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [2.2.1-alpha.0](https://github.com/baloise/ui-library/compare/v2.2.0-alpha.2...v2.2.1-alpha.0) (2021-01-04)
+
+
+
+## 2.1.1 (2020-12-23)
+
+**Note:** Version bump only for package root
+
+
+
+
+
+## [2.1.1](https://github.com/baloise/ui-library/compare/v2.1.0...v2.1.1) (2020-12-23)
+
+
+### Bug Fixes
+
+* **utils:** refactor utils ([de8c21d](https://github.com/baloise/ui-library/commit/de8c21d55c9641dc1ff89192011c6bd7bdaf9903))
+
+
+
+
+
+# [2.1.0](https://github.com/baloise/ui-library/compare/v2.0.3...v2.1.0) (2020-12-22)
+
+
+### Features
+
+* **utils:** add claim-number util ([669683f](https://github.com/baloise/ui-library/commit/669683fd6bf1b27d27da0ab5f6983c4f1669c01e))
+
+
+
+
+
 ## [2.0.3](https://github.com/baloise/ui-library/compare/v2.0.2...v2.0.3) (2020-12-22)
 
 **Note:** Version bump only for package root