+++ packages/testing/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.4](https://github.com/baloise/ui-library/compare/v2.3.3...v2.3.4) (2021-01-04)
+
+**Note:** Version bump only for package @baloise/ui-library-testing
+
+
+
+
+
+## [2.3.3](https://github.com/baloise/ui-library/compare/v2.3.2...v2.3.3) (2021-01-04)
+
+**Note:** Version bump only for package @baloise/ui-library-testing
+
+
+
+
+
+## [2.3.2](https://github.com/baloise/ui-library/compare/v2.3.1...v2.3.2) (2021-01-04)
+
+**Note:** Version bump only for package @baloise/ui-library-testing
+
+
+
+
+
+## [2.3.1](https://github.com/baloise/ui-library/compare/v2.3.0...v2.3.1) (2021-01-04)
+
+**Note:** Version bump only for package @baloise/ui-library-testing
+
+
+
+
+
+## [2.2.1-alpha.0](https://github.com/baloise/ui-library/compare/v2.2.0-alpha.2...v2.2.1-alpha.0) (2021-01-04)
+
+
+
+## 2.1.1 (2020-12-23)
+
+**Note:** Version bump only for package @baloise/ui-library-testing
+
+
+
+
+
+## [2.1.1](https://github.com/baloise/ui-library/compare/v2.1.0...v2.1.1) (2020-12-23)
+
+**Note:** Version bump only for package @baloise/ui-library-testing
+
+
+
+
+
+# [2.1.0](https://github.com/baloise/ui-library/compare/v2.0.3...v2.1.0) (2020-12-22)
+
+**Note:** Version bump only for package @baloise/ui-library-testing
+
+
+
+
+
 ## [2.0.3](https://github.com/baloise/ui-library/compare/v2.0.2...v2.0.3) (2020-12-22)
 
 **Note:** Version bump only for package @baloise/ui-library-testing